Update README.md
What is Customer Churn?
Customer churn is defined as when customers or subscribers discontinue doing business with a
firm or service.
Customers in the telecom industry can choose from a variety of service providers and actively
switch from one to the next. The telecommunications business has an annual churn rate of 15-25
percent in this highly competitive market.
Individualized customer retention is tough because most firms have a large number of customers
and can't afford to devote much time to each of them. The costs would be too great, outweighing
the additional revenue. However, if a corporation could forecast which customers are likely to
leave ahead of time, it could focus customer retention efforts only on these "high risk" clients. The
ultimate goal is to expand its coverage area and retrieve more customers loyalty. The core to
succeed in this market lies in the customer itself.
Customer churn is a critical metric because it is much less expensive to retain existing customers
than it is to acquire new customers.
To detect early signs of potential churn, one must first develop a holistic view of the customers
and their interactions across numerous channels. As a result, by addressing churn, these
businesses may not only preserve their market position, but also grow and thrive. More
customers they have in their network, the lower the cost of initiation and the larger the profit. As
a result, the company's key focus for success is reducing client attrition and implementing
effective retention strategy.
Business Aspects
Customers usually do not decide to switch to another competitor instantly, but rather over a
period (especially applicable to high-value customers). In churn prediction, we assume that there
are three phases of customer lifecycle:
The ‘good’ phase: In this phase, the customer is happy with the service and behaves as usual.
The ‘action’ phase: The customer experience starts to sore in this phase, for e.g. he/she gets a
compelling offer from a competitor, faces unjust charges, becomes unhappy with service quality
etc. In this phase, the customer usually shows different behavior than the ‘good’ months. Also, it
is crucial to identify high-churn-risk customers in this phase, since some corrective actions can be
taken at this point
The ‘churn’ phase: In this phase, the customer is said to have churned. You define churn based on
this phase. Also, it is important to note that at the time of prediction, this data is not available to
you for prediction. Thus, after tagging churn as 1/0 based on this phase, you discard all data
corresponding to this phase.